+++ lib/mongoid/relations/many.rb
@@ -40,9 +40,13 @@ module Mongoid
       #
       # @since 2.0.0.beta.1
       def create(attributes = nil, type = nil, &block)
-        doc = build(attributes, type, &block)
-        base.persisted? ? doc.save : raise_unsaved(doc)
-        doc
+        if attributes.is_a?(::Array)
+          attributes.map { |attrs| create(attrs, type, &block) }
+        else
+          doc = build(attributes, type, &block)
+          base.persisted? ? doc.save : raise_unsaved(doc)
+          doc
+        end
       end
 
       # Creates a new document on the references many relation. This will
@@ -66,9 +70,13 @@ module Mongoid
       #
       # @since 2.0.0.beta.1
       def create!(attributes = nil, type = nil, &block)
-        doc = build(attributes, type, &block)
-        base.persisted? ? doc.save! : raise_unsaved(doc)
-        doc
+        if attributes.is_a?(::Array)
+          attributes.map { |attrs| create!(attrs, type, &block) }
+        else
+          doc = build(attributes, type, &block)
+          base.persisted? ? doc.save! : raise_unsaved(doc)
+          doc
+        end
       end
 
       # Find the first document given the conditions, or creates a new document